# medical-dictation-backend/docs/api/report-service-openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PhiAccessCreate": {
        "additionalProperties": false,
        "properties": {
          "reason_code": {
            "enum": [
              "patient_complaint",
              "legal_request",
              "billing_dispute",
              "quality_review",
              "care_continuity",
              "data_correction",
              "other"
            ],
            "title": "Reason Code",
            "type": "string"
          },
          "reason_note": {
            "default": "",
            "maxLength": 2000,
            "title": "Reason Note",
            "type": "string"
          },
          "reauth_ticket": {
            "description": "Single-use ticket from POST /auth/reauth (auth-service).",
            "maxLength": 512,
            "minLength": 1,
            "title": "Reauth Ticket",
            "type": "string"
          },
          "resource_id": {
            "description": "The report to open.",
            "format": "uuid",
            "title": "Resource Id",
            "type": "string"
          }
        },
        "required": [
          "resource_id",
          "reason_code",
          "reauth_ticket"
        ],
        "title": "PhiAccessCreate",
        "type": "object"
      },
      "PhiAccessList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PhiAccessOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PhiAccessList",
        "type": "object"
      },
      "PhiAccessOut": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "granted_at": {
            "format": "date-time",
            "title": "Granted At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "patient_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "reason_note": {
            "title": "Reason Note",
            "type": "string"
          },
          "requested_by": {
            "format": "uuid",
            "title": "Requested By",
            "type": "string"
          },
          "resource_id": {
            "format": "uuid",
            "title": "Resource Id",
            "type": "string"
          },
          "resource_kind": {
            "title": "Resource Kind",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "revoked_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked By"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "use_count": {
            "title": "Use Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "requested_by",
          "resource_kind",
          "resource_id",
          "patient_id",
          "reason_code",
          "reason_note",
          "status",
          "granted_at",
          "expires_at",
          "revoked_at",
          "revoked_by",
          "use_count",
          "last_used_at"
        ],
        "title": "PhiAccessOut",
        "type": "object"
      },
...
      "ReasonCatalog": {
        "additionalProperties": false,
        "properties": {
          "grant_ttl_minutes": {
            "title": "Grant Ttl Minutes",
            "type": "integer"
          },
          "note_min_chars": {
            "title": "Note Min Chars",
            "type": "integer"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonOption"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "reasons",
          "grant_ttl_minutes",
          "note_min_chars"
        ],
        "title": "ReasonCatalog",
        "type": "object"
      },
      "ReasonOption": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "label_en": {
            "title": "Label En",
            "type": "string"
          },
          "label_uk": {
            "title": "Label Uk",
            "type": "string"
          },
          "requires_note": {
            "title": "Requires Note",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "label_uk",
          "label_en",
          "requires_note"
        ],
        "title": "ReasonOption",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Sprint-06 templates + sprint-08 reports / versions / diff / search.",
    "title": "Report Service",
    "version": "0.8.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/phi-access-requests": {
      "get": {
        "operationId": "list_requests_v1_phi_access_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "resource_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Id"
            }
          },
          {
            "in": "query",
            "name": "requested_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested By"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhiAccessList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oversight: who broke glass, on what, why, and how often they used it.",
        "tags": [
          "phi-access"
        ]
      },
      "post": {
        "operationId": "create_request_v1_phi_access_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhiAccessCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhiAccessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Break glass on one report: reason + password re-entry \u2192 a time-limited grant.",
        "tags": [
          "phi-access"
        ]
      }
    },
    "/v1/phi-access-requests/reasons": {
      "get": {
        "operationId": "list_reasons_v1_phi_access_requests_reasons_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The break-glass reason vocabulary and grant window.",
        "tags": [
          "phi-access"
        ]
      }
    },
    "/v1/phi-access-requests/{grant_id}/revoke": {
      "post": {
        "operationId": "revoke_request_v1_phi_access_requests__grant_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "grant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Grant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhiAccessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close an open grant early.",
        "tags": [
          "phi-access"
        ]
      }
    },
...
  }
}